Implemented animation by start/stop/period & fixed frame rate bug.
Frame rate for GIFs was not obeying the URL parameter.  It is now.

Animation didn't support start/stop/period.  This meant that for fine
resolution / large datasets, the time string (generated and sent back to
the client from the WMS) was too long to be contained as a URL
parameter.  By supporting period strings and replacing the generated
timestrings with these, the problem goes away.