Improved interfaces
Added ContinuousDomainDataset and make appropriate datasets implement
it.  This means we don't check for an abstract class in WmsServlet.

That in turn allows us to create datasets with a continuous domain 
which don't have to subclass AbstractContinuousDomainDataset, meaning
that a FeatureIndexer is not a prerequisite of such a dataset.

These changes were made to make it more amenable to a WMS extension in
another project.

Also moved SimpleLayerNameMapper to edal-graphics - it is a simple
implementation independent of the XML stuff in edal-xml-catalogue where
it previously resided.